+++ weconnect/auth/we_charge_session.py
@@ -229,6 +229,8 @@ class WeChargeSession(VWWebSession):
             token_url = add_params_to_uri(token_url, urlParams)
 
             tokenResponse = self.get(token_url, headers=loginHeadersForm, allow_redirects=False, access_type=AccessType.ID)
+            if tokenResponse.status_code != requests.codes['ok']:
+                raise TemporaryAuthentificationError('Token could not be fetched due to temporary WeConnect failure: {tokenResponse.status_code}')
 
             self.parseFromBody(tokenResponse.text)